tienda_forestal_mvc_vue/backend/requirements.txt
flask
flask-cors
flask-mysqldb

